feat: Refactor directory structure for domain separation (#73)
