+++ src/main/java/com/example/ssauc/user/cash/controller/CashController.java
@@ -1,0 +1,15 @@
+package com.example.ssauc.user.cash.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@Controller
+@RequestMapping("cash")
+public class CashController {
+
+    @GetMapping("/cash")  // SSAUC 머니 페이지로 이동
+    public String cashPage() {
+        return "/cash/cash";
+    }
+}
