+++ src/main/java/com/example/ssauc/user/cash/entity/Charge.java
@@ -1,4 +1,4 @@
-package com.example.ssauc.user.mypage.entity;
+package com.example.ssauc.user.cash.entity;
 
 
 import com.example.ssauc.user.login.entity.Users;
