+++ src/main/java/com/example/ssauc/user/cash/repository/ChargeRepository.java
@@ -1,6 +1,6 @@
-package com.example.ssauc.user.mypage.repository;
+package com.example.ssauc.user.cash.repository;
 
-import com.example.ssauc.user.mypage.entity.Charge;
+import com.example.ssauc.user.cash.entity.Charge;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ChargeRepository extends JpaRepository<Charge, Long> {
