+++ src/main/java/com/example/ssauc/user/cash/repository/WithdrawRepository.java
@@ -1,6 +1,6 @@
-package com.example.ssauc.user.mypage.repository;
+package com.example.ssauc.user.cash.repository;
 
-import com.example.ssauc.user.mypage.entity.Withdraw;
+import com.example.ssauc.user.cash.entity.Withdraw;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface WithdrawRepository extends JpaRepository<Withdraw, Long> {
