+++ src/main/java/com/example/ssauc/user/login/entity/Users.java
@@ -9,14 +9,13 @@ import com.example.ssauc.user.contact.entity.Board;
 import com.example.ssauc.user.main.entity.Notification;
 import com.example.ssauc.user.main.entity.ProductLike;
 import com.example.ssauc.user.main.entity.RecentlyViewed;
-import com.example.ssauc.user.mypage.entity.Charge;
+import com.example.ssauc.user.cash.entity.Charge;
 import com.example.ssauc.user.mypage.entity.ReputationHistory;
 import com.example.ssauc.user.mypage.entity.UserActivity;
-import com.example.ssauc.user.mypage.entity.Withdraw;
+import com.example.ssauc.user.cash.entity.Withdraw;
 import com.example.ssauc.user.order.entity.Orders;
 import com.example.ssauc.user.pay.entity.Payment;
 import com.example.ssauc.user.pay.entity.Review;
-import com.example.ssauc.user.product.entity.Category;
 import com.example.ssauc.user.product.entity.Product;
 import jakarta.persistence.*;
 import lombok.*;
@@ -62,6 +61,8 @@ public class Users {
     private LocalDateTime createdAt;
     private LocalDateTime updatedAt;
     private LocalDateTime lastLogin;
+
+    private Long cash;
 
     // 🔹 추가: username과 password만 받는 생성자
     public Users(String userName, String password) {
