+++ src/main/resources/templates/history/ban.html
